requirements.txt
a2wsgi==1.10.10
aiosmtplib==4.0.2
aiosqlite==0.21.0
alembic==1.16.5
annotated-types==0.7.0
anyio==4.11.0
apache-airflow==3.1.0
apache-airflow-providers-common-compat==1.7.4
apache-airflow-providers-common-io==1.6.3
apache-airflow-providers-common-sql==1.28.1
apache-airflow-providers-smtp==2.2.1
apache-airflow-providers-standard==1.8.0
apache-airflow-providers-trino==6.3.3
argcomplete==3.6.2
asgiref==3.9.2
attrs==25.3.0
cadwyn==5.4.4
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
clickhouse-driver==0.2.9
colorlog==6.9.0
cron-descriptor==2.0.6
croniter==6.0.0
cryptography==46.0.0
deprecated==1.2.18
dill==0.4.0
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.117.1
fastapi-cli==0.0.13
fsspec==2025.9.0
googleapis-common-protos==1.70.0
greenback==1.2.1
greenlet==3.2.4
grpcio==1.75.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
importlib-metadata==8.7.0
itsdangerous==2.2.0
jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
lazy-object-proxy==1.12.0
libcst==1.8.4
linkify-it-py==2.0.3
lockfile==0.12.2
mako==1.3.10
markdown-it-py==4.0.0
markupsafe==3.0.2
mdurl==0.1.2
methodtools==0.4.7
more-itertools==10.8.0
msgspec==0.19.0
natsort==8.4.0
opentelemetry-api==1.37.0
opentelemetry-exporter-otlp==1.37.0
opentelemetry-exporter-otlp-proto-common==1.37.0
opentelemetry-exporter-otlp-proto-grpc==1.37.0
opentelemetry-exporter-otlp-proto-http==1.37.0
opentelemetry-proto==1.37.0
opentelemetry-sdk==1.37.0
opentelemetry-semantic-conventions==0.58b0
outcome==1.3.0.post0
packaging==25.0
pathlib-abc==0.5.1
pathspec==0.12.1
pendulum==3.1.0
pluggy==1.6.0
protobuf==6.32.1
psutil==7.1.0
pycparser==2.23
pydantic==2.11.9
pydantic-core==2.33.2
pygments==2.19.2
pygtrie==2.5.0
pyjwt==2.10.1
python-daemon==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
python-slugify==8.0.4
pytz==2025.2
pyyaml==6.0.2
pyyaml-ft==8.0.0
referencing==0.36.2
requests==2.32.5
retryhttp==1.3.3
rich==13.9.4
rich-argparse==1.7.1
rich-toolkit==0.15.1
rpds-py==0.27.1
ruff==0.14.0
setproctitle==1.3.7
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sqlalchemy==2.0.43
sqlalchemy-jsonfield==1.0.2
sqlalchemy-utils==0.42.0
sqlparse==0.5.3
starlette==0.48.0
structlog==25.4.0
style==1.1.0
svcs==25.1.0
tabulate==0.9.0
tenacity==9.1.2
termcolor==3.1.0
text-unidecode==1.3
typer==0.19.2
types-requests==2.32.4.20250913
typing-extensions==4.15.0
typing-inspection==0.4.1
tzdata==2025.2
tzlocal==5.2
uc-micro-py==1.0.3
universal-pathlib==0.2.6
update==0.0.1
urllib3==2.5.0
uuid6==2025.0.1
uvicorn==0.37.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
wirerope==1.0.0
wrapt==1.17.3
zipp==3.23.0
